Moves dex trade types to common
